--- a/cpp/src/arrow/testing/gtest_util.cc
+++ b/cpp/src/arrow/testing/gtest_util.cc
@@ -68,8 +68,7 @@ void AssertTsEqual(const T& expected, const T& actual) {
   }
 }
 
-void AssertArraysEqual(const Array& expected, const Array& actual,
-                       bool verbose) {
+void AssertArraysEqual(const Array& expected, const Array& actual, bool verbose) {
   std::stringstream diff;
   if (!expected.Equals(actual, EqualOptions().diff_sink(&diff))) {
     if (verbose) {