--- a/cpp/src/parquet/arrow/reader_internal.cc
+++ b/cpp/src/parquet/arrow/reader_internal.cc
@@ -564,7 +564,7 @@ Status GetOriginSchema(const std::shared_ptr<const KeyValueMetadata>& metadata,
 
   if (metadata->size() > 1) {
     // Copy the metadata without the schema key
-    auto new_metadata = ::arrow::key_value_metadata({});
+    auto new_metadata = ::arrow::key_value_metadata({}, {});
     new_metadata->reserve(metadata->size() - 1);
     for (int64_t i = 0; i < metadata->size(); ++i) {
       if (i == schema_index) continue;