… cpp/CMakeLists.txt
… cpp/src/parquet/arrow/arrow-reader-writer-test.cc
… cpp/src/parquet/column_writer-test.cc
… cpp/src/parquet/encoding-test.cc
… cpp/src/parquet/encoding.h
… cpp/src/parquet/metadata.cc
… cpp/src/parquet/properties.h
… cpp/src/parquet/statistics-test.cc
… cpp/src/parquet/statistics.cc
… cpp/src/parquet/statistics.h
… cpp/src/parquet/types.h
… python/pyarrow/_parquet.pxd
… python/pyarrow/_parquet.pyx
… python/pyarrow/tests/test_parquet.py