There's a lot going of interconnected pieces in this patch, so let me try to unpack:

* Refactor `TypedColumnWriterImpl::WriteBatch/WriteBatchSpaced` to utilize more common code and be more readable
* Add `TypedEncoder<T>::Put(const arrow::Array&)` and implement for BYTE_ARRAY so avoid having to first create `ByteArray*` as before. This should improve write performance for regular binary data -- I will do some benchmarks to measure by how much
* Add `TypedStatistics<T>::Update(const arrow::Array&)` and implement for BYTE_ARRAY. This is necessary to be able to update the statistics given directly-inserted Arrow data without serialization
* Implement `PutDictionary` and `PutIndices` methods on `DictEncoder`. `PutDictionary` is only implemented for BYTE_ARRAY but can be easily generalized to more types (we should open a follow up JIRA for this)
* Implement internal `TypedColumnWriterImpl::WriteArrowDictionary` that writes dictionary values and indices directly into a DictEncoder. This circumvents the dictionary page size checks so that we will continue to call `PutIndices` until a new dictionary is encountered or some non-dictionary data is written. Note that in master, dictionary encoding is turned off as soon as a threshold in dictionary size is reached, which is by default 1MB. So if you want to preserve exactly the original dictionary values (e.g. if you are roundtripping DictionaryArray, R factor, or pandas.Categorical), then we have to step around this threshold check in this narrow case.
* Add `ArrowWriterProperties::store_schema()` option which stores the Arrow schema used to create a Parquet file in a special `ARROW:schema` key in the metadata, so that we can detect that a column was originally DictionaryArray. This option is off by default, but enabled in the Python bindings. We can always make it the default in the future

I think that's most things. One end result of this is that `arrow::DictionaryArray` types from C++ and `pandas.Categorical` types coerced from pandas with string dictionary values will be accurately preserved end-to-end. With a little more work (which I think can be done in a follow up PR) we can support the rest of the Parquet physical types.

This was a fairly ugly project and I've doubtlessly left some messes around that we should clean up, but perhaps in follow up patches. 

I'll post some benchmarks later to assess improvements in read and write performance. In the case of writing dictionary-encoded strings the boost should be significant.