@wesm 2019-08-14
Looks like I have an ASAN failure and some compiler warnings on Windows. Will investigate tomorrow