@wesm 2019-08-14
OK I think I've fixed the issues that popped up. I know there's a lot to review here, so we may have to leave many of the refinements to follow up work