@wesm 2019-08-15
See before and after benchmarks:

https://gist.github.com/wesm/b9e917b46536676d2ed829bc4458ba3d

Summary, for this synthetic example

* Write performance of DictionaryArray improved 4x from 800ms to 200ms
* Peak memory use when writing drops from ~700MB to ~50MB because a cast to dense is no longer required
* Read performance (comparing direct dictionary read vs. read-as-dense) improved 5x from ~500ms to ~100ms
* Writing of regular BinaryArray as BYTE_ARRAY improved by ~10% (see "dense_data" write at the bottom of each benchmark)