@wesm 2019-08-15
The Appveyor failure in the R build seems legitimate, this is annoying:

```
C:/projects/arrow/cpp/src/parquet/arrow/reader_internal.cc: In function 'arrow::Status parquet::arrow::GetOriginSchema(const std::shared_ptr<const arrow::KeyValueMetadata>&, std::shared_ptr<const arrow::KeyValueMetadata>*, std::shared_ptr<arrow::Schema>*)':
C:/projects/arrow/cpp/src/parquet/arrow/reader_internal.cc:567:55: error: converting to 'const std::unordered_map<std::basic_string<char>, std::basic_string<char> >' from initializer list would use explicit constructor 'std::unordered_map<_Key, _Tp, _Hash, _Pred, _Alloc>::unordered_map(std::unordered_map<_Key, _Tp, _Hash, _Pred, _Alloc>::size_type, const hasher&, const key_equal&, const allocator_type&) [with _Key = std::basic_string<char>; _Tp = std::basic_string<char>; _Hash = std::hash<std::basic_string<char> >; _Pred = std::equal_to<std::basic_string<char> >; _Alloc = std::allocator<std::pair<const std::basic_string<char>, std::basic_string<char> > >; std::unordered_map<_Key, _Tp, _Hash, _Pred, _Alloc>::size_type = long long unsigned int; std::unordered_map<_Key, _Tp, _Hash, _Pred, _Alloc>::hasher = std::hash<std::basic_string<char> >; std::unordered_map<_Key, _Tp, _Hash, _Pred, _Alloc>::key_equal = std::equal_to<std::basic_string<char> >; std::unordered_map<_Key, _Tp, _Hash, _Pred, _Alloc>::allocator_type = std::allocator<std::pair<const std::basic_string<char>, std::basic_string<char> > >]'
     auto new_metadata = ::arrow::key_value_metadata({});
                                                       ^
make[2]: *** [src/parquet/CMakeFiles/parquet_static.dir/build.make:90: src/parquet/CMakeFiles/parquet_static.dir/arrow/reader_internal.cc.obj] Error 1
make[2]: *** Waiting for unfinished jobs....
C:/projects/arrow/cpp/src/parquet/arrow/writer.cc: In function 'arrow::Status parquet::arrow::GetSchemaMetadata(const arrow::Schema&, arrow::MemoryPool*, const parquet::ArrowWriterProperties&, std::shared_ptr<const arrow::KeyValueMetadata>*)':
C:/projects/arrow/cpp/src/parquet/arrow/writer.cc:569:44: error: converting to 'const std::unordered_map<std::basic_string<char>, std::basic_string<char> >' from initializer list would use explicit constructor 'std::unordered_map<_Key, _Tp, _Hash, _Pred, _Alloc>::unordered_map(std::unordered_map<_Key, _Tp, _Hash, _Pred, _Alloc>::size_type, const hasher&, const key_equal&, const allocator_type&) [with _Key = std::basic_string<char>; _Tp = std::basic_string<char>; _Hash = std::hash<std::basic_string<char> >; _Pred = std::equal_to<std::basic_string<char> >; _Alloc = std::allocator<std::pair<const std::basic_string<char>, std::basic_string<char> > >; std::unordered_map<_Key, _Tp, _Hash, _Pred, _Alloc>::size_type = long long unsigned int; std::unordered_map<_Key, _Tp, _Hash, _Pred, _Alloc>::hasher = std::hash<std::basic_string<char> >; std::unordered_map<_Key, _Tp, _Hash, _Pred, _Alloc>::key_equal = std::equal_to<std::basic_string<char> >; std::unordered_map<_Key, _Tp, _Hash, _Pred, _Alloc>::allocator_type = std::allocator<std::pair<const std::basic_string<char>, std::basic_string<char> > >]'
     result = ::arrow::key_value_metadata({});
                                            ^
make[2]: *** [src/parquet/CMakeFiles/parquet_static.dir/build.make:116: src/parquet/CMakeFiles/parquet_static.dir/arrow/writer.cc.obj] Error 1
make[1]: *** [CMakeFiles/Makefile2:1265: src/parquet/CMakeFiles/parquet_static.dir/all] Error 2
make: *** [Makefile:141: all] Error 2
==> ERROR: A failure occurred in build().
    Aborting...
```