@bkietz 2019-08-15
https://github.com/apache/arrow/pull/5077/files#diff-66bf5c14b0dc83dceaebb3fe0ddbb55cR567
I'd recommend building new_metadata as an unordered_map then calling key_value_metadata to wrap it after it's built.