@wesm 2019-08-16
Thank you for the code review!