@wesm 2019-08-16
Travis CI: https://travis-ci.org/wesm/arrow/builds/572728109
Appveyor: https://ci.appveyor.com/project/ApacheSoftwareFoundation/arrow/builds/26746111